Rename module path resolver
