+++ lib/app/src/common/mod.rs
@@ -1,5 +1,5 @@
 pub mod dependency_serializer;
+pub mod file_path_resolver;
 pub mod interface_serializer;
 pub mod module_id_calculator;
-pub mod module_path_resolver;
 pub mod package_id_calculator;
